requirements.txt
Django>=3.1
django-cassiopeia>=2.0.0
